Added root index to kong.yml - master

+++ docker-kong.sh
@@ -24,16 +24,21 @@ if [[ $(sudo docker inspect -f '{{.State.Running}}' $CONTAINER_NAME) = "true" ]]
         echo "Stopped previous Container: $CONTAINER_NAME"
 fi
 
-echo "REMOVING OLD CONTAINER: $CONTAINER_NAME"
-sudo docker rm $CONTAINER_NAME
+if [[ $(sudo docker inspect -f '{{.State.Running}}' $CONTAINER_NAME) = "false" ]]; then
+        sudo docker rm $CONTAINER_NAME
+fi 
+
+# if [[ $(sudo docker ps -a | grep $CONTAINER_NAME) = "true" ]]; then
+#       echo "REMOVING OLD CONTAINER: $CONTAINER_NAME"
+#       sudo docker rm $CONTAINER_NAME
+# fi
+
 
 echo "Starting Docker image name: $DOCKER_IMAGE"
 
 echo $DOCKER_PWD | sudo docker login -u $DOCKER_LOGIN --password-stdin
 
 echo "Will run docker run command"
-
-
 sudo docker run  -d  --network=host --restart always \
         -e "KONG_DATABASE=off" \
         -e "KOND_DECLARATIVE_CONFIG=/app/kong.yml" \
@@ -42,7 +47,10 @@ sudo docker run  -d  --network=host --restart always \
         -e "KONG_PROXY_ERROR_LOG=/dev/stderr" \
         -e "KONG_ADMIN_ERROR_LOG=/dev/stderr" \
         -e "KONG_ADMIN_LISTEN=0.0.0.0:8001, 0.0.0.0:8444 ssl" \
+        -p 8000:8000 \
+        -p 8001:8001 \
         --name $CONTAINER_NAME $DOCKER_IMAGE
+
 
 echo "Command run complete"
